+++ MvRAdaptiveCards/ArgumentCompleters/New-CardBadge.ps1
@@ -1,0 +1,10 @@
+[Diagnostics.CodeAnalysis.SuppressMessageAttribute('PSReviewUnusedParameter', '')]
+param()
+Register-ArgumentCompleter -CommandName New-CardBadge -ParameterName icon -ScriptBlock {
+    param ( $commandName,
+        $parameterName,
+        $wordToComplete,
+        $commandAst,
+        $fakeBoundParameters )
+    $_AdaptiveCardIconCollection | Where-Object { $_ -like "$wordToComplete*" } | ForEach-Object { [System.Management.Automation.CompletionResult]::new($_, $_, 'ParameterValue', $_) }
+}